shared: cast to float in Com_Scale

+++ code/qcommon/q_shared.h
@@ -775,7 +775,7 @@ void PerpendicularVector( vec3_t dst, const vec3_t src );
 //=============================================
 
 float Com_Clamp( float min, float max, float value );
-#define Com_Scale(x, min, max, newmin, newmax)	((newmin) + ((newmax)-(newmin)) * ((x)-(min)) / ((max)-(min)))
+#define Com_Scale(x, min, max, newmin, newmax)	((newmin) + ((newmax)-(newmin)) * (float)((x)-(min)) / (float)((max)-(min)))
 #define Com_Sign(x)	(((x) > 0) - ((x) < 0))
 
 char	*COM_SkipPath( char *pathname );
